+++ CHANGELOG.md
@@ -5,6 +5,24 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.4.0] - 2024-11-17
+
+### Fixes
+
+* Correct an issue with the assember's addressing-mode matcher
+* Correct some issues that could come up in disassembly
+* Various forms of `RUN` can end an Applesoft relocation flow
+
+### New Features
+
+* Support converting data to code in Merlin source files
+
+### New Behaviors
+
+* Prevent excessively long workspace scans
+    - Limit the directory count and recursion depth
+    - Skip `build`, `node_modules`, and `target` directories
+
 ## [3.3.3] - 2024-10-26
 
 ### Fixes